fix: split service.py to clear the AI-readiness file-size hard cap
abicheck/service.py grew to 2011 lines, tripping the 2000-line hard
limit and failing test_no_hard_file_size_violations and
test_main_returns_zero_on_clean_tree. Move render_output and its
helpers into a new leaf module, abicheck/service_render.py, following
the existing service_scan.py extraction pattern, and re-export them
from service.py for API compatibility.
